refactor: organized serve
